Combine modules (#34)
* move cache module to data module

* remove cache models and mappers

* sort products by position from db

* delete testUtils module

* move data module to domain module

* remove domain and data models with associated mappers

* move domain module to core

* move presentation module to shopping_list

* move core module to shopping_list

+++ shopping_list/src/main/java/com/zizohanto/android/tobuy/shopping_list/presentation/mappers/ProductModelMapper.kt
@@ -1,8 +1,7 @@
 package com.zizohanto.android.tobuy.shopping_list.presentation.mappers
 
-import com.zizohanto.android.tobuy.domain.models.Product
-import com.zizohanto.android.tobuy.presentation.mapper.ModelMapper
 import com.zizohanto.android.tobuy.shopping_list.presentation.models.ProductsViewItem.ProductModel
+import com.zizohanto.android.tobuy.shoppinglist.sq.Product
 import javax.inject.Inject
 
 class ProductModelMapper @Inject constructor() : ModelMapper<ProductModel, Product> {
@@ -14,7 +13,7 @@ class ProductModelMapper @Inject constructor() : ModelMapper<ProductModel, Produ
                 shoppingListId,
                 name,
                 price,
-                position
+                position.toInt()
             )
         }
     }
@@ -26,7 +25,7 @@ class ProductModelMapper @Inject constructor() : ModelMapper<ProductModel, Produ
                 shoppingListId,
                 name,
                 price,
-                position
+                position.toLong()
             )
         }
     }
